Creates Scaffold folder and moves ScreenScaffolded, TopBar and BottomBar
