+++ Start/app/src/main/java/com/esei/grvidal/nighttime/scaffold/ScaffoldedScreen.kt
@@ -1,4 +1,4 @@
-package com.esei.grvidal.nighttime
+package com.esei.grvidal.nighttime.scaffold
 
 import androidx.compose.foundation.Text
 import androidx.compose.foundation.clickable
@@ -8,14 +8,16 @@ import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.esei.grvidal.nighttime.CustomDialog
 import com.esei.grvidal.nighttime.data.City
 
 /**
  * Screen with the Scaffolded format, with the topBar with the button to change the city
  * and the bottomBar with the navigation buttons
  *
- * @param navController Controller of the navigation
- * @param items list of the selectable navigation buttons
+ * @param modifier Modifier of the Screen
+ * @param bottomBar  content to be shown on the bottom of the screen
+ * @param topBar content to be shown on the top of the screen
  * @param content content to be shown on the center of the screen
  */
 @Composable
@@ -42,6 +44,13 @@ fun ScreenScaffolded(
     }
 }
 
+/**
+ * City Dialog stateful function, it will set the parameters to show CityDialog
+ *
+ * @param cityDialog Boolean to show or hide cityDialog
+ * @param items List of items to be Shown
+ * @param setCityDialog
+ */
 @Composable
 fun CityDialogConstructor(
     cityDialog: Boolean,
@@ -53,7 +62,7 @@ fun CityDialogConstructor(
         CustomDialog(onClose = { setCityDialog(false) }) {
             CityDialog(
                 items = items,
-                editCity = { city ->
+                onItemClick = { city ->
                     setCityId(city.id,city.name)
                     setCityDialog(false)
                 }
@@ -67,12 +76,12 @@ fun CityDialogConstructor(
  * Dialog with the cities that can be selected
  *
  * @param items list of the cities
- * @param editCity setter of the selected city
+ * @param onItemClick Unit to do when an item is selected
  */
 @Composable
 fun CityDialog(
     items: List<City>,
-    editCity: (City) -> Unit
+    onItemClick: (City) -> Unit
 ) {
     LazyColumnFor(items = items) {
         Row(
@@ -81,7 +90,7 @@ fun CityDialog(
             Surface(
                 modifier = Modifier
                     .preferredWidth(120.dp)
-                    .clickable(onClick = { editCity(it) } ),
+                    .clickable(onClick = { onItemClick(it) } ),
                 color = MaterialTheme.colors.background
             ) {
                 Text(
