+++ Start/app/src/main/java/com/esei/grvidal/nighttime/scaffold/TopBar.kt
@@ -1,4 +1,4 @@
-package com.esei.grvidal.nighttime
+package com.esei.grvidal.nighttime.scaffold
 
 import androidx.compose.foundation.Text
 import androidx.compose.foundation.clickable
@@ -16,6 +16,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.esei.grvidal.nighttime.R
 import java.util.*
 
 @Composable
